# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get list of tasks",
                "operationId": "6a93d54730cde509d735e538ab6800f6",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TaskResource"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginatorLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginatorMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a task",
                "operationId": "9b481c0ed5fdfa1a592d70ba39a348ee",
                "requestBody": {
                    "description": "New Task",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Unprocessable Content"
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get a task",
                "operationId": "fd3677b5b24695d7b801715869b1357c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update a task",
                "description": "Update the specified resource in storage.",
                "operationId": "85377ff0e935f093dd75ea6886923dbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "New data for a task",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Unprocessable Content"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete a task",
                "operationId": "5ed5dfd4c645e9398f4282969bfe6478",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResource"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "PaginatorMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "from": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "example": "http://localhost/api/entities?page=1"
                                },
                                "label": {
                                    "example": "&laquo; Previous"
                                },
                                "active": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "path": {
                        "example": "http://localhost/api/entities"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "to": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "PaginatorLinks": {
                "properties": {
                    "first": {
                        "example": "http://localhost/api/entities?page=1"
                    },
                    "last": {
                        "example": "http://localhost/api/entities?page=5"
                    },
                    "prev": {
                        "example": "http://localhost/api/entities?page=1"
                    },
                    "next": {
                        "example": "http://localhost/api/entities?page=3"
                    }
                },
                "type": "object"
            },
            "TaskRequest": {
                "required": [
                    "title",
                    "description",
                    "status",
                    "executor_id"
                ],
                "properties": {
                    "title": {
                        "example": "Title"
                    },
                    "description": {
                        "example": "Description"
                    },
                    "status": {
                        "format": "string",
                        "example": "planned"
                    }
                },
                "type": "object"
            },
            "MessageResource": {
                "properties": {
                    "message": {
                        "example": "Text message."
                    }
                },
                "type": "object"
            },
            "TaskResource": {
                "properties": {
                    "id": {
                        "example": "2"
                    },
                    "title": {
                        "example": "Title"
                    },
                    "description": {
                        "example": "Description"
                    },
                    "status": {
                        "format": "string",
                        "example": "planned"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "Tasks"
        }
    ]
}